+++ build.gradle.kts
@@ -52,12 +52,4 @@ allprojects {
         configFile = rootProject.file("resources/checkstyle-config.xml")
         configDirectory.set(rootProject.file("resources"))
     }
-
-    // EdcRuntimeExtension uses this to determine the runtime classpath of the module to run.
-    tasks.register("printClasspath") {
-        doLast {
-            println(sourceSets["main"].runtimeClasspath.asPath)
-        }
-    }
-    
 }
