+++ extensions/common/gcp/gcp-core/src/main/java/org/eclipse/edc/gcp/common/GcpAccessToken.java
@@ -16,7 +16,7 @@ package org.eclipse.edc.gcp.common;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import org.eclipse.edc.connector.transfer.spi.types.SecretToken;
+import org.eclipse.edc.connector.controlplane.transfer.spi.types.SecretToken;
 
 @JsonTypeName("dataspaceconnector:gcptoken")
 public class GcpAccessToken implements SecretToken {
